Update Test.md
This is a line of test code for testing skyhooh.It is worth noting that the code in skyhook is not rigorous, and it simplifies many complicated processes. I just had a whim.That line of code omits a lot of processes, which may just be a record of inspiration.If you want to use it, go ahead. I don't think many people want to use it either.